Reorganized/renamed observer interfaces into new package.
Like with the observable interfaces, they have also been renamed as
I<ObserverType>Observer.
