+++ Final_Project/src/main/java/connect4Menu/control/Connect4SettingsController.java
@@ -14,7 +14,6 @@ import javax.swing.event.ChangeListener;
 import connect4Menu.exceptions.InvalidTimerTimeException;
 import connect4Menu.model.Connect4SettingsModel;
 import connect4Menu.model.player.IPlayerColors;
-import connect4Menu.model.player.Player;
 import connect4Menu.view.Connect4SettingsView;
 import mvcinterfaces.MenuController;
 
@@ -88,6 +87,13 @@ public class Connect4SettingsController implements MenuController {
 			parseForTimerTime(textField.getText());	
 		}
 		
+		/**
+		 * Parses the String given to see if it's an integer and if it's within range. If so, it'll change the value in the model.
+		 * Otherwise, it will throw an exception. 
+		 * @param timerTextFieldText text from timerTextField.
+		 * @return String of errorLabel.
+		 * @author Khuram C.
+		 */
 		public String parseForTimerTime(String timerTextFieldText){
 			try {
 				int timerTextFieldValue = Integer.parseInt(timerTextFieldText);
@@ -123,6 +129,12 @@ public class Connect4SettingsController implements MenuController {
 			changeColor((JComboBox<IPlayerColors>) e.getSource());
 		}
 		
+		/**
+		 * Changes the color of the player associated with the playerBox.
+		 * @param comboBox to get color from.
+		 * @return boolean detailing success.
+		 * @author Khuram C.
+		 */
 		public boolean changeColor(JComboBox<IPlayerColors> comboBox) {
 			IPlayerColors color = (IPlayerColors) comboBox.getSelectedItem();
 			model.changePlayerColor(color.getAllowedPlayer(), color);
@@ -142,6 +154,11 @@ public class Connect4SettingsController implements MenuController {
 			startConnect4Round();
 		}
 		
+		/**
+		 * Starts a round of Connect4 by calling initiate in the gameController.
+		 * @return boolean detailing success.
+		 * @author Khuram C.
+		 */
 		public boolean startConnect4Round() {
 			Connect4GameController gameController = new Connect4GameController(model);
 			gameController.initiate();
