+++ Final_Project/src/main/java/connect4Menu/model/observableinterfaces/IEndGameObservable.java
@@ -1,11 +1,11 @@
 package connect4Menu.model.observableinterfaces;
 
 import connect4Menu.model.player.Player;
-import connect4Menu.view.ObserverEndGame;
+import connect4Menu.view.observerinterfaces.IEndGameObserver;
 
 public interface IEndGameObservable {
 
-	public void registerEndGameObserver(ObserverEndGame o);
+	public void registerEndGameObserver(IEndGameObserver o);
 
 	public void notifyEndGameObservers(Player p);
