+++ Final_Project/src/main/java/connect4Menu/model/observableinterfaces/IInvalidColsObservable.java
@@ -1,10 +1,10 @@
 package connect4Menu.model.observableinterfaces;
 
-import connect4Menu.view.ObserverInvalidCols;
+import connect4Menu.view.observerinterfaces.IInvalidColsObserver;
 
 public interface IInvalidColsObservable {
 
-	public void registerInvalidColObserver(ObserverInvalidCols o);
+	public void registerInvalidColObserver(IInvalidColsObserver o);
 
 	public void notifyInvalidColObservers(int colNum);
