+++ Final_Project/src/main/java/connect4Menu/model/observableinterfaces/ISquarePlayedObservable.java
@@ -1,11 +1,11 @@
 package connect4Menu.model.observableinterfaces;
 
 import connect4Menu.model.player.Player;
-import connect4Menu.view.ObserverSquarePlayed;
+import connect4Menu.view.observerinterfaces.ISquarePlayedObserver;
 
 public interface ISquarePlayedObservable {
 
-	public void registerSquarePlayedObserver(ObserverSquarePlayed o);
+	public void registerSquarePlayedObserver(ISquarePlayedObserver o);
 
 	public void notifySquarePlayedObservers(Player p);
