+++ Final_Project/src/main/java/connect4Menu/model/observableinterfaces/IStartedTurnObservable.java
@@ -1,11 +1,11 @@
 package connect4Menu.model.observableinterfaces;
 
 import connect4Menu.model.player.Player;
-import connect4Menu.view.ObserverStartedTurn;
+import connect4Menu.view.observerinterfaces.IStartedTurnObserver;
 
 public interface IStartedTurnObservable {
 
-	public void registerStartedTurnObserver(ObserverStartedTurn o);
+	public void registerStartedTurnObserver(IStartedTurnObserver o);
 
 	public void notifyStartedTurnObservers(Player p);
 }
