--- Final_Project/src/main/java/connect4Menu/view/ObserverInvalidCols.java
@@ -1,7 +1,0 @@
-package connect4Menu.view;
-
-public interface ObserverInvalidCols {
-
-	public void updateButton(int invalidColNum);
-
-}
