+++ Final_Project/src/main/java/connect4Menu/view/observerinterfaces/IInvalidColsObserver.java
@@ -1,0 +1,7 @@
+package connect4Menu.view.observerinterfaces;
+
+public interface IInvalidColsObserver {
+
+	public void updateButton(int invalidColNum);
+
+}
