+++ Final_Project/src/main/java/connect4Menu/view/observerinterfaces/IStartedTurnObserver.java
@@ -1,8 +1,8 @@
-package connect4Menu.view;
+package connect4Menu.view.observerinterfaces;
 
 import connect4Menu.model.player.Player;
 
-public interface ObserverStartedTurn {
+public interface IStartedTurnObserver {
 
 	public void updatePlayerTurnText(Player p);
